+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix shell completion for arguments.
+- Allow empty module for rsync.
 
 ## [1.0.0] - 2021-10-30
 ### Added